manual-tests/requirements.txt
psutil==5.9.6
PyYAML==6.0.1
